\name{default.barplot.theme}
\alias{default.barplot.theme}
\title{
Default barplot theme.
}
\description{
A theme function that defines a set of default values for barplots created by this package.  This theme is based
on theme_bw() and modifies a few parameters.
}
\usage{
default.barplot.theme(base_size = 24, base_family = "Helvetica", xaxis.angle = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{base_size}{Base font size for the figure (default: 24)}
  \item{base_family}{Base font family to use for the figure (default: Helveticar)}
  \item{xaxis.angle}{Change the x-axis tick labels}
}
\author{
Richard de Borja <richard.deborja@sickkids.ca>
}
\keyword{ barplot }
\keyword{ theme }
