\name{default.boxplot.theme}
\alias{default.boxplot.theme}
\title{Default boxplot theme}
\description{
This function provides the set of default "theme" values for boxplots.  The original base theme is theme_bw().
}
\usage{
default.boxplot.theme(base_size = 30, base_family = "Helvetica")
}
\arguments{
  \item{base_size}{Base font size for the plot (default: 30)}
  \item{base_family}{Base font family for the plot (default: Helvetica)}
}
\value{Returns a "theme" object to be used with ggplot2}
\references{
http://docs.ggplot2.org/dev/vignettes/themes.html
}
\author{
Richard de Borja <richard.deborja@sickkids.ca>
}
\keyword{ plotting.general }
\keyword{ theme }
\keyword{ ggplot2 }
\keyword{ boxplot }
\keyword{ default }