\name{default.histogram.theme}
\alias{default.histogram.theme}
\title{Default theme for a histogram}
\description{A default histogram ggplot2 theme for publication ready plots}
\usage{
default.histogram.theme(base_size = 30, base_family = "Helvetica")
}
\arguments{
  \item{base_size}{Base font size}
  \item{base_family}{Base font family}
}
\author{
Richard de Borja <richard.deborja@sickkids.ca>
}
\keyword{ histogram }
\keyword{ theme }
\keyword{ ggplot2 }
