\name{plotting.general-package}
\alias{plotting.general-package}
\alias{plotting.general}
\docType{package}
\title{Plotting functions using ggplot2}
\description{A set of functions wrapping ggplot2 to generate publication ready figures.}
\details{
\tabular{ll}{
Package: \tab plotting.general\cr
Type: \tab Package\cr
Version: \tab 0.12\cr
Date: \tab 2014-11-11\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes \cr
}
create.barplot()
create.boxplot()
create.heatmap()
create.histogram()
create.scatterplot()
create.stripplot()
write.plot()
}
\author{
Author: Richard de Borja
Maintainer: Richard de Borja <richard.deborja@sickkids.ca>
}
\references{
Paul Murrell (2011). R Graphics (2nd ed.). Boca Raton, FL: CRC Press 
}
\keyword{ plotting }
\keyword{ ggplot2 }
