\name{plotting.general-package}
\alias{plotting.general-package}
\alias{plotting.general}
\docType{package}
\title{Plotting functions using ggplot2}
\description{A set of functions wrapping ggplot2 to generate publication ready figures.}
\details{
\tabular{ll}{
Package: \tab plotting.general\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-02-04\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes \cr
}
create.barplot()
}
\author{
Author: Richard de Borja
Maintainer: Richard de Borja <richard.deborja@sickkids.ca>
}
\keyword{ ggplot2 }
